added a way to store intermediate resutls, so we do not need to download
all the time from the web. note: this loses factor levers as we save in
csv.
divided the work in three files:
1) definitions of the functions.
2) load the data. this part can be executed each day to get updated
data.
3) graphing: make the grapfs for the required lists of countries.
note that plm is used oly to make the diffs.
we can now inttroduce lags when we make the extra variables.
this is important for the recovered-deatsh vs confirmed graphs.

On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   Covid19.csv
	new file:   Covid19_US.csv
	modified:   Covid19_days.csv
	new file:   Covid19_non_US.csv
	modified:   Graphs.R
	new file:   USpop2019.csv
	new file:   USstatespop2019.csv
	modified:   definitions2.R
	modified:   deprecated.R
	new file:   geo.location.US.csv
	modified:   geo.location.csv
	new file:   loadData.R
	new file:   population.csv
	modified:   wip.R